src/test/java/use_case/friends_list_user_story/friend_profile_friends_list/FriendProfileFriendsListInteractorTest.java
//package use_case.friends_list_user_story.friend_profile_friends_list;
//
//import org.junit.Test;
//import static org.junit.Assert.*;
//
//public class FriendProfileFriendsListInteractorTest {
//
//    @Test
//    public void testSwitchToFriendProfileView() {
//        // Arrange
//        FriendProfileFriendsListOutputBoundary presenter = new FriendProfileFriendsListOutputBoundary() {
//            @Override
//            public void switchToFriendProfileView(String selectedFriendName, String password) {
//                assertEquals("TestSelectedFriend", selectedFriendName);
//                assertEquals("TestPassword", password);
//            }
//        };
//
//        FriendProfileFriendsListInteractor friendProfileFriendsListInteractor = new FriendProfileFriendsListInteractor(presenter);
//
//        // Act
//        friendProfileFriendsListInteractor.switchToFriendProfileView("TestSelectedFriend", "TestPassword");
//
//        // Assert
//        // The assertion is done within the presenter's switchToFriendProfileView method.
//    }
//
//    @Test(expected = NullPointerException.class)
//    public void testConstructorWithNullPresenter_Refactored() {
//        new FriendProfileFriendsListInteractor(null);
//    }
//
//    @Test(expected = NullPointerException.class)
//    public void testSwitchToFriendProfileViewWithNullPresenter_Refactored() {
//        FriendProfileFriendsListInteractor friendProfileFriendsListInteractor = new FriendProfileFriendsListInteractor(null);
//
//        // Act
//        friendProfileFriendsListInteractor.switchToFriendProfileView("TestSelectedFriend", "TestPassword");
//    }
//}